Remove Feature.ENABLE_CBM and all related code
