--- app/data/data-chat-read-timestamp/data-chat-read-timestamp-public/src/main/graphql/QueryChatLatestMessageTimestamps.graphql
@@ -1,7 +1,0 @@
-query ChatLatestMessageTimestamps {
-  chat(until: null) {
-    messages {
-      sentAt
-    }
-  }
-}
