+++ app/data/data-chat-read-timestamp/data-chat-read-timestamp-public/src/main/kotlin/com/hedvig/android/data/chat/read/timestamp/ChatLastMessageReadRepository.kt
@@ -2,37 +2,24 @@ package com.hedvig.android.data.chat.read.timestamp
 
 import arrow.core.merge
 import arrow.core.raise.either
-import arrow.core.toNonEmptyListOrNull
 import com.apollographql.apollo.ApolloClient
 import com.apollographql.apollo.cache.normalized.FetchPolicy
 import com.apollographql.apollo.cache.normalized.fetchPolicy
 import com.benasher44.uuid.Uuid
 import com.hedvig.android.apollo.safeExecute
 import com.hedvig.android.data.chat.database.ConversationDao
-import com.hedvig.android.featureflags.FeatureManager
-import com.hedvig.android.featureflags.flags.Feature
 import kotlin.time.Duration.Companion.seconds
 import kotlinx.coroutines.currentCoroutineContext
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.flow.transformLatest
+import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.isActive
 import kotlinx.datetime.Instant
 import octopus.CbmChatLatestMessageTimestampsQuery
 import octopus.CbmChatLatestMessageTimestampsQuery.Data.CurrentMember
-import octopus.ChatLatestMessageTimestampsQuery
 import octopus.type.ChatMessageSender.MEMBER
 
 interface ChatLastMessageReadRepository {
-  /**
-   * Reports the timestamp of the newest message which the member has seen. Ignores it if an already newer timestamp
-   * has been reported.
-   *
-   * TODO remove when CBM is enabled. With it, the stored timestamp happens from the RemoteMediator directly to the DB
-   */
-  suspend fun storeLatestReadTimestamp(timestamp: Instant)
-
   /**
    * Returns [true] if the newest message is newer than the last message which the member has already seen.
    * Continuously queries the latest messages to emit the most up-to-date value.
@@ -41,69 +28,37 @@ interface ChatLastMessageReadRepository {
 }
 
 internal class ChatLastMessageReadRepositoryImpl(
-  private val chatMessageTimestampStorage: ChatMessageTimestampStorage,
   private val apolloClient: ApolloClient,
-  private val featureManager: FeatureManager,
   private val conversationDao: ConversationDao,
 ) : ChatLastMessageReadRepository {
-  override suspend fun storeLatestReadTimestamp(timestamp: Instant) {
-    if (featureManager.isFeatureEnabled(Feature.ENABLE_CBM).first()) return
-    val existingTimestamp = chatMessageTimestampStorage.getLatestReadTimestamp()
-    if (existingTimestamp != null && existingTimestamp > timestamp) {
-      return
-    }
-    chatMessageTimestampStorage.setLatestReadTimestamp(timestamp)
-  }
-
   override fun isNewestMessageNewerThanLastReadTimestamp(): Flow<Boolean> {
-    return featureManager.isFeatureEnabled(Feature.ENABLE_CBM).transformLatest { isCbmEnabled ->
+    return flow {
       while (currentCoroutineContext().isActive) {
-        if (isCbmEnabled) {
-          val isNewestMessageNewerThanLastReadTimestamp = either {
-            val currentMember = apolloClient.query(CbmChatLatestMessageTimestampsQuery())
-              .fetchPolicy(FetchPolicy.NetworkOnly)
-              .safeExecute()
-              .bind()
-              .currentMember
-            val allNewestHedvigMessages = buildList {
-              add(currentMember.legacyConversation?.toNewestHedvigMessage())
-              addAll(currentMember.conversations.map { it.toNewestHedvigMessage() })
-            }.filterNotNull()
-            val backendTimestamps = allNewestHedvigMessages.associate { it.conversationId to it.lastMessageTimestamp }
-            val databaseTimestamps: Map<String, Instant> = conversationDao
-              .getLatestTimestamps(allNewestHedvigMessages.map { Uuid.fromString(it.conversationId) })
-              .associate { it.id.toString() to it.lastMessageReadTimestamp }
-            for ((id, backendTimestamp) in backendTimestamps) {
-              val databaseTimestamp = databaseTimestamps[id]
-              if (databaseTimestamp != null && databaseTimestamp < backendTimestamp) {
-                return@either true
-              }
-            }
-            false
-          }
-            .mapLeft { false }
-            .merge()
-          emit(isNewestMessageNewerThanLastReadTimestamp)
-        } else {
-          val lastReadMessageTimestamp: Instant? = chatMessageTimestampStorage.getLatestReadTimestamp()
-          val messages = apolloClient
-            .query(ChatLatestMessageTimestampsQuery())
+        val isNewestMessageNewerThanLastReadTimestamp = either {
+          val currentMember = apolloClient.query(CbmChatLatestMessageTimestampsQuery())
             .fetchPolicy(FetchPolicy.NetworkOnly)
             .safeExecute()
-            .getOrNull()
-            ?.chat
-            ?.messages
-            ?.toNonEmptyListOrNull()
-          if (messages == null) {
-            emit(false)
-          } else if (lastReadMessageTimestamp == null) {
-            // If there are existing messages, but we have seen none of them, there always is an unread message
-            emit(true)
-          } else {
-            val newestChatMessageTimestamp = messages.maxOf { it.sentAt }
-            emit(lastReadMessageTimestamp < newestChatMessageTimestamp)
+            .bind()
+            .currentMember
+          val allNewestHedvigMessages = buildList {
+            add(currentMember.legacyConversation?.toNewestHedvigMessage())
+            addAll(currentMember.conversations.map { it.toNewestHedvigMessage() })
+          }.filterNotNull()
+          val backendTimestamps = allNewestHedvigMessages.associate { it.conversationId to it.lastMessageTimestamp }
+          val databaseTimestamps: Map<String, Instant> = conversationDao
+            .getLatestTimestamps(allNewestHedvigMessages.map { Uuid.fromString(it.conversationId) })
+            .associate { it.id.toString() to it.lastMessageReadTimestamp }
+          for ((id, backendTimestamp) in backendTimestamps) {
+            val databaseTimestamp = databaseTimestamps[id]
+            if (databaseTimestamp != null && databaseTimestamp < backendTimestamp) {
+              return@either true
+            }
           }
+          false
         }
+          .mapLeft { false }
+          .merge()
+        emit(isNewestMessageNewerThanLastReadTimestamp)
         delay(10.seconds)
       }
     }
