--- app/feature/feature-chat/src/main/graphql/FragmentMessageFragment.graphql
@@ -1,12 +1,0 @@
-fragment MessageFragment on ChatMessage {
-  id
-  sender
-  sentAt
-  ... on ChatMessageFile {
-    signedUrl
-    mimeType
-  }
-  ... on ChatMessageText {
-    text
-  }
-}
