--- app/feature/feature-chat/src/main/graphql/MutationChatSendFile.graphql
@@ -1,10 +1,0 @@
-mutation ChatSendFile($uploadToken: ID!, $context: ChatMessageContext) {
-  chatSendFile(input: { uploadToken: $uploadToken, context: $context } ) {
-    message {
-      ...MessageFragment
-    }
-    error {
-      message
-    }
-  }
-}
