--- app/feature/feature-chat/src/main/graphql/MutationChatSendMessage.graphql
@@ -1,10 +1,0 @@
-mutation ChatSendMessage($text: String!, $context: ChatMessageContext) {
-  chatSendText(input: { text: $text, context: $context }) {
-    message {
-      ...MessageFragment
-    }
-    error {
-      message
-    }
-  }
-}
