--- app/feature/feature-chat/src/main/graphql/QueryChatMessages.graphql
@@ -1,11 +1,0 @@
-query ChatMessages($until: Instant) {
-  chat(until: $until) {
-    id
-    hasNext
-    nextUntil
-    messages {
-      ...MessageFragment
-    }
-    bannerText
-  }
-}
