+++ app/feature/feature-chat/src/main/kotlin/com/hedvig/android/feature/chat/CbmChatViewModel.kt
@@ -1,4 +1,4 @@
-package com.hedvig.android.feature.chat.cbm
+package com.hedvig.android.feature.chat
 
 import android.net.Uri
 import androidx.compose.runtime.Composable
@@ -28,16 +28,19 @@ import com.hedvig.android.data.chat.database.ChatDao
 import com.hedvig.android.data.chat.database.ChatMessageEntity
 import com.hedvig.android.data.chat.database.ConversationDao
 import com.hedvig.android.data.chat.database.RemoteKeyDao
-import com.hedvig.android.feature.chat.cbm.ConversationInfo.Info
-import com.hedvig.android.feature.chat.cbm.ConversationInfo.NoConversation
-import com.hedvig.android.feature.chat.cbm.ConversationInfoStatus.Failed
-import com.hedvig.android.feature.chat.cbm.ConversationInfoStatus.Initializing
-import com.hedvig.android.feature.chat.cbm.ConversationInfoStatus.Loaded
-import com.hedvig.android.feature.chat.cbm.model.CbmChatMessage
-import com.hedvig.android.feature.chat.cbm.model.Sender
-import com.hedvig.android.feature.chat.cbm.model.toChatMessage
-import com.hedvig.android.feature.chat.cbm.model.toLatestChatMessage
-import com.hedvig.android.feature.chat.cbm.paging.ChatRemoteMediator
+import com.hedvig.android.feature.chat.ConversationInfoStatus.Failed
+import com.hedvig.android.feature.chat.ConversationInfoStatus.Initializing
+import com.hedvig.android.feature.chat.ConversationInfoStatus.Loaded
+import com.hedvig.android.feature.chat.data.BannerText
+import com.hedvig.android.feature.chat.data.CbmChatRepository
+import com.hedvig.android.feature.chat.data.ConversationInfo
+import com.hedvig.android.feature.chat.data.ConversationInfo.Info
+import com.hedvig.android.feature.chat.data.ConversationInfo.NoConversation
+import com.hedvig.android.feature.chat.model.CbmChatMessage
+import com.hedvig.android.feature.chat.model.Sender
+import com.hedvig.android.feature.chat.model.toChatMessage
+import com.hedvig.android.feature.chat.model.toLatestChatMessage
+import com.hedvig.android.feature.chat.paging.ChatRemoteMediator
 import com.hedvig.android.logger.logcat
 import com.hedvig.android.molecule.android.MoleculeViewModel
 import com.hedvig.android.molecule.public.MoleculePresenter
