+++ app/feature/feature-chat/src/main/kotlin/com/hedvig/android/feature/chat/di/ChatModule.kt
@@ -9,20 +9,14 @@ import com.hedvig.android.data.chat.database.AppDatabase
 import com.hedvig.android.data.chat.database.ChatDao
 import com.hedvig.android.data.chat.database.ConversationDao
 import com.hedvig.android.data.chat.database.RemoteKeyDao
-import com.hedvig.android.data.chat.read.timestamp.ChatLastMessageReadRepository
-import com.hedvig.android.feature.chat.ChatViewModel
-import com.hedvig.android.feature.chat.cbm.CbmChatRepositoryImpl
-import com.hedvig.android.feature.chat.cbm.CbmChatViewModel
-import com.hedvig.android.feature.chat.cbm.data.CbmChatRepositoryDemo
-import com.hedvig.android.feature.chat.cbm.data.GetAllConversationsUseCase
-import com.hedvig.android.feature.chat.cbm.data.GetAllConversationsUseCaseImpl
-import com.hedvig.android.feature.chat.cbm.data.GetCbmChatRepositoryProvider
-import com.hedvig.android.feature.chat.cbm.inbox.InboxViewModel
+import com.hedvig.android.feature.chat.CbmChatViewModel
 import com.hedvig.android.feature.chat.data.BotServiceService
-import com.hedvig.android.feature.chat.data.ChatRepositoryDemo
-import com.hedvig.android.feature.chat.data.ChatRepositoryImpl
-import com.hedvig.android.feature.chat.data.GetChatRepositoryProvider
-import com.hedvig.android.navigation.core.AppDestination
+import com.hedvig.android.feature.chat.data.CbmChatRepositoryDemo
+import com.hedvig.android.feature.chat.data.CbmChatRepositoryImpl
+import com.hedvig.android.feature.chat.data.GetAllConversationsUseCase
+import com.hedvig.android.feature.chat.data.GetAllConversationsUseCaseImpl
+import com.hedvig.android.feature.chat.data.GetCbmChatRepositoryProvider
+import com.hedvig.android.feature.chat.inbox.InboxViewModel
 import kotlinx.datetime.Clock
 import kotlinx.serialization.json.Json
 import okhttp3.MediaType.Companion.toMediaType
@@ -33,33 +27,6 @@ import retrofit2.Retrofit
 import retrofit2.converter.kotlinx.serialization.asConverterFactory
 
 val chatModule = module {
-  viewModel<ChatViewModel> { parametersHolder ->
-    val chatContext = parametersHolder.getOrNull<AppDestination.Chat.ChatContext>()
-    ChatViewModel(
-      chatRepository = get<GetChatRepositoryProvider>(),
-      clock = get<Clock>(),
-      chatContext = chatContext,
-    )
-  }
-  single<ChatRepositoryImpl> {
-    ChatRepositoryImpl(
-      apolloClient = get<ApolloClient>(),
-      botServiceService = get<BotServiceService>(),
-      fileService = get<FileService>(),
-      chatLastMessageReadRepository = get<ChatLastMessageReadRepository>(),
-    )
-  }
-  single<ChatRepositoryDemo> {
-    ChatRepositoryDemo(get<Clock>())
-  }
-  single<GetChatRepositoryProvider> {
-    GetChatRepositoryProvider(
-      demoManager = get<DemoManager>(),
-      demoImpl = get<ChatRepositoryDemo>(),
-      prodImpl = get<ChatRepositoryImpl>(),
-    )
-  }
-
   single<BotServiceService> {
     val retrofit = Retrofit
       .Builder()
@@ -71,7 +38,6 @@ val chatModule = module {
     retrofit.create(BotServiceService::class.java)
   }
 
-  // cbm
   single<GetAllConversationsUseCase> {
     GetAllConversationsUseCaseImpl(get<ApolloClient>(), get<ConversationDao>())
   }
