--- app/feature/feature-chat/src/main/kotlin/com/hedvig/android/feature/chat/model/ChatMessagesResult.kt
@@ -1,10 +1,0 @@
-package com.hedvig.android.feature.chat.model
-
-import kotlinx.datetime.Instant
-
-data class ChatMessagesResult(
-  val messages: List<ChatMessage>,
-  val nextUntil: Instant,
-  val hasNext: Boolean,
-  val informationMessage: String?,
-)
