+++ app/feature/feature-chat/src/main/kotlin/com/hedvig/android/feature/chat/model/InboxConversation.kt
@@ -1,4 +1,4 @@
-package com.hedvig.android.feature.chat.cbm.model
+package com.hedvig.android.feature.chat.model
 
 import kotlinx.datetime.Instant
