+++ app/feature/feature-chat/src/main/kotlin/com/hedvig/android/feature/chat/model/Sender.kt
@@ -1,4 +1,4 @@
-package com.hedvig.android.feature.chat.cbm.model
+package com.hedvig.android.feature.chat.model
 
 import com.hedvig.android.data.chat.database.ChatMessageEntity
 import octopus.type.ChatMessageSender
