+++ app/feature/feature-claim-details/src/main/graphql/QueryClaims.graphql
@@ -1,8 +1,8 @@
-query Claims($isCbmEnabled: Boolean!) {
+query Claims {
   currentMember {
     claims {
       ...ClaimFragment
-      conversation @include(if: $isCbmEnabled) {
+      conversation {
         id
       }
     }
