+++ app/featureflags/feature-flags-public/src/main/kotlin/com/hedvig/android/featureflags/flags/Feature.kt
@@ -16,5 +16,4 @@ enum class Feature(
   ),
   EDIT_COINSURED("Let member edit co insured"),
   HELP_CENTER("Enable the help center screens"),
-  ENABLE_CBM("Use the new conversational based messaging system, and the new paging backend API"),
 }
